Moved component to material-ui-lab
