+++ packages/material-ui-lab/src/Slider/index.js
